fs: move sync status to common
Issue: KBFS-4012
