+++ shared/fs/common/index.js
@@ -12,3 +12,4 @@ export {default as OpenChat} from './open-chat'
 export {default as NewFolder} from './new-folder'
 export {default as UploadButton} from './upload-button'
 export {default as FolderViewFilter} from './folder-view-filter'
+export {default as SyncStatus} from './sync-status-container'
